+++ app/javascript/react/lib/constants.js
